New blower mount and clean up
Flipped the blower motor mount so that it is reversed. This puts the
fans buisness end closer to the hot end. It does have the fan intake
pointing at the extruder but hopefully is far enough away as to not
matter.

Cleaned up testing and reference model stl files that don't need to be
in the repository.