Reverting Motor Mount
Reverting the motor mount to no longer include the mounting holes for
the bltouch. It made the mount too brittle in places and also you have
to take apart the hot end to get adjust the sensor which isn't
ideal. Going to try and mount the sensor in a different way.